refactor: moved DeobfuscateTask & DecompileTask to evaluated/ directory
Took 5 minutes
